requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
gunicorn

pydantic==2.13.0
pydantic-settings==2.13.1

python-dotenv==1.2.2
SQLAlchemy==2.0.49
requests==2.33.1
pyodbc==5.3.0
azure-storage-blob==12.28.0
boto3==1.42.88
python-multipart==0.0.26
email-validator==2.3.0